#338
Refactor code to create another sub base class "AccreditationBaseController.cs" as per PR comments.
